Skeleton playback enabled - alpha
-Ability to playback skeleton roughly implemented; will need lots of
work

-Added “record mode” to differentiate between when user is recording a
dance and when user is playing back a dance (dance meaning skeleton).
Initiated by pressing “shift + control” on the login screen.

TO FIX:
-NullPointerException is created when Dancecraft is run without Kinect
being plugged in.

-Skeleton playback is not correct; needs debugging.  Delay function
inserted in the playBack function does not allow for any other input
while skeleton is being redrawn.  App must be stopped and restarted.